comments, python worker results clean
